requirements.txt
aiohttp==3.7.3
aiohttp-socks==0.5.5
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
chardet==4.0.0
colorama==0.4.4
colorlog==4.6.2
croniter==0.3.36
graia-application-mirai==0.11.0
graia-broadcast==0.5.0
humanfriendly==9.0
idna==2.10
iterwrapper==0.1.4
lxml==4.6.2
multidict==5.1.0
natsort==7.1.0
Pillow==8.0.1
pydantic==1.7.3
pyreadline==2.1
python-dateutil==2.8.1
python-socks==1.1.1
regex==2020.11.13
six==1.15.0
soupsieve==2.0.1
tqdm==4.54.1
typing-extensions==3.7.4.3
yarl==1.6.3
openpyxl~=3.0.5
APScheduler~=3.6.3
pandas~=1.1.5
matplotlib~=3.3.3
xpinyin~=0.7.4